+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/MapboxCarApp.kt
@@ -1,10 +1,11 @@
 package com.mapbox.androidauto
 
-import android.app.Application
 import com.mapbox.androidauto.navigation.location.CarAppLocation
 import com.mapbox.androidauto.navigation.location.impl.CarAppLocationImpl
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
+import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
 import com.mapbox.navigation.ui.app.internal.SharedApp
 import com.mapbox.navigation.ui.voice.internal.MapboxAudioGuidance
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,10 +14,10 @@ import kotlinx.coroutines.flow.StateFlow
 /**
  * The entry point for your Mapbox Android Auto app.
  */
-@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
-object MapboxCarApp {
-
+object MapboxCarApp : MapboxNavigationObserver {
     private val carAppStateFlow = MutableStateFlow<CarAppState>(FreeDriveState)
+    @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
+    private val carAppLocation: CarAppLocation = CarAppLocationImpl()
 
     /**
      * Attach observers to the CarAppState to determine which view to show.
@@ -26,8 +27,7 @@ object MapboxCarApp {
     /**
      * Location service available to the car and app.
      */
-    fun carAppLocationService(): CarAppLocation =
-        MapboxNavigationApp.getObserver(CarAppLocation::class)
+    fun carAppLocationService(): CarAppLocation = carAppLocation
 
     /**
      * Audio guidance service available to the car and app.
@@ -42,13 +42,13 @@ object MapboxCarApp {
         carAppStateFlow.value = carAppState
     }
 
-    /**
-     * Setup android auto from your [Application.onCreate]
-     *
-     * @param application used to detect when activities are foregrounded
-     */
-    fun setup(application: Application) {
-        SharedApp.setup(application)
-        MapboxNavigationApp.registerObserver(CarAppLocationImpl())
+    override fun onAttached(mapboxNavigation: MapboxNavigation) {
+        // TODO add after 2.8.0-rc.1
+//        MapboxNavigationApp.registerObserver(SharedApp)
+        MapboxNavigationApp.registerObserver(carAppLocation)
+    }
+
+    override fun onDetached(mapboxNavigation: MapboxNavigation) {
+        MapboxNavigationApp.unregisterObserver(carAppLocation)
     }
 }