+++ libnavui-app/src/main/java/com/mapbox/navigation/ui/app/internal/SharedApp.kt
@@ -2,6 +2,7 @@ package com.mapbox.navigation.ui.app.internal
 
 import android.content.Context
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.internal.extensions.attachCreated
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
@@ -23,9 +24,7 @@ import com.mapbox.navigation.ui.voice.internal.impl.MapboxAudioGuidanceImpl
 import java.util.concurrent.atomic.AtomicBoolean
 
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
-object SharedApp {
-    private var isSetup = false
-
+object SharedApp : MapboxNavigationObserver {
     val store = Store()
     val state get() = store.state.value
 
@@ -42,6 +41,10 @@ object SharedApp {
     val destinationStateController = DestinationStateController(store)
     val routeStateController = RouteStateController(store)
     val routePreviewStateController = RoutePreviewStateController(store)
+    private val stateResetController = StateResetController(store, ignoreTripSessionUpdates)
+    private val routeAlternativeComponent = RouteAlternativeComponent {
+        RouteAlternativeComponentImpl(store)
+    }
     private val navigationObservers: Array<MapboxNavigationObserver> = arrayOf(
         routeStateController,
         cameraStateController,
@@ -51,26 +54,14 @@ object SharedApp {
         routePreviewStateController,
         audioGuidanceStateController,
         tripSessionStarterStateController,
+        stateResetController,
+        routeAlternativeComponent,
     )
 
-    @JvmOverloads
-    fun setup(
-        context: Context,
-        audioGuidance: MapboxAudioGuidance? = null,
-        routeAlternativeContract: RouteAlternativeContract? = null
-    ) {
-        if (isSetup) return
-        isSetup = true
-
-        MapboxNavigationApp.registerObserver(StateResetController(store, ignoreTripSessionUpdates))
-        MapboxNavigationApp.registerObserver(
-            RouteAlternativeComponent {
-                routeAlternativeContract ?: RouteAlternativeComponentImpl(store)
-            }
-        )
-        MapboxNavigationApp.lifecycleOwner.attachCreated(*navigationObservers)
-        MapboxNavigationApp.registerObserver(audioGuidance ?: defaultAudioGuidance(context))
-    }
+    /**
+     * Requires the [Context] so will be initialized when attached.
+     */
+    private lateinit var mapboxAudioGuidance: MapboxAudioGuidance
 
     fun tripSessionTransaction(updateSession: () -> Unit) {
         // Any changes to MapboxNavigation TripSession should be done within `tripSessionTransaction { }` block.
@@ -85,6 +76,18 @@ object SharedApp {
         return MapboxAudioGuidanceImpl.create(context).also {
             it.dataStoreOwner = NavigationDataStoreOwner(context, DEFAULT_DATA_STORE_NAME)
         }
+    }
+
+    override fun onAttached(mapboxNavigation: MapboxNavigation) {
+        val context = mapboxNavigation.navigationOptions.applicationContext
+        mapboxAudioGuidance = defaultAudioGuidance(context)
+        MapboxNavigationApp.registerObserver(mapboxAudioGuidance)
+        MapboxNavigationApp.lifecycleOwner.attachCreated(*navigationObservers)
+    }
+
+    override fun onDetached(mapboxNavigation: MapboxNavigation) {
+        navigationObservers.forEach { MapboxNavigationApp.unregisterObserver(it) }
+        MapboxNavigationApp.unregisterObserver(mapboxAudioGuidance)
     }
 
     private const val DEFAULT_DATA_STORE_NAME = "mapbox_navigation_preferences"