+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/QaTestApplication.kt
@@ -1,13 +1,5 @@
 package com.mapbox.navigation.qa_test_app
 
 import android.app.Application
-import com.mapbox.androidauto.MapboxCarApp
 
-class QaTestApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        MapboxCarApp.setup(this)
-    }
-}
+class QaTestApplication : Application()